requirements.txt
tensorflow~=2.8.0
gym~=0.22.0
opencv-python~=4.5.5.62
PyYAML~=6.0
numpy~=1.22.2
mariadb~=1.0.10